requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0
PyMuPDF==1.23.8
sentence-transformers==2.7.0
transformers==4.37.0
huggingface_hub==0.20.3
faiss-cpu==1.7.4
pinecone-client==3.2.2
google-generativeai==0.7.2
numpy==1.24.3
pandas==1.5.3
scikit-learn==1.3.2
python-multipart==0.0.6
aiofiles==23.2.0
httpx==0.25.2
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
